+++ oci/monitoring_alarm_resource.go
@@ -1,0 +1,583 @@
+// Copyright (c) 2017, Oracle and/or its affiliates. All rights reserved.
+
+package provider
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/hashicorp/terraform/helper/schema"
+
+	oci_common "github.com/oracle/oci-go-sdk/common"
+	oci_monitoring "github.com/oracle/oci-go-sdk/monitoring"
+)
+
+func AlarmResource() *schema.Resource {
+	return &schema.Resource{
+		Importer: &schema.ResourceImporter{
+			State: schema.ImportStatePassthrough,
+		},
+		Timeouts: DefaultTimeout,
+		Create:   createAlarm,
+		Read:     readAlarm,
+		Update:   updateAlarm,
+		Delete:   deleteAlarm,
+		Schema: map[string]*schema.Schema{
+			// Required
+			"compartment_id": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"destinations": {
+				Type:     schema.TypeList,
+				Required: true,
+				Elem: &schema.Schema{
+					Type: schema.TypeString,
+				},
+			},
+			"display_name": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"is_enabled": {
+				Type:     schema.TypeBool,
+				Required: true,
+			},
+			"metric_compartment_id": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"namespace": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"query": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+			"severity": {
+				Type:     schema.TypeString,
+				Required: true,
+			},
+
+			// Optional
+			"body": {
+				Type:     schema.TypeString,
+				Optional: true,
+				Computed: true,
+			},
+			"defined_tags": {
+				Type:             schema.TypeMap,
+				Optional:         true,
+				Computed:         true,
+				DiffSuppressFunc: definedTagsDiffSuppressFunction,
+				Elem:             schema.TypeString,
+			},
+			"freeform_tags": {
+				Type:     schema.TypeMap,
+				Optional: true,
+				Computed: true,
+				Elem:     schema.TypeString,
+			},
+			"metric_compartment_id_in_subtree": {
+				Type:     schema.TypeBool,
+				Optional: true,
+				Computed: true,
+			},
+			"pending_duration": {
+				Type:     schema.TypeString,
+				Optional: true,
+				Computed: true,
+			},
+			"repeat_notification_duration": {
+				Type:     schema.TypeString,
+				Optional: true,
+				Computed: true,
+			},
+			"resolution": {
+				Type:     schema.TypeString,
+				Optional: true,
+				Computed: true,
+			},
+			"suppression": {
+				Type:     schema.TypeList,
+				Optional: true,
+				Computed: true,
+				MaxItems: 1,
+				MinItems: 1,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						// Required
+						"time_suppress_from": {
+							Type:             schema.TypeString,
+							Required:         true,
+							DiffSuppressFunc: timeDiffSuppressFunction,
+						},
+						"time_suppress_until": {
+							Type:             schema.TypeString,
+							Required:         true,
+							DiffSuppressFunc: timeDiffSuppressFunction,
+						},
+
+						// Optional
+						"description": {
+							Type:     schema.TypeString,
+							Optional: true,
+							Computed: true,
+						},
+
+						// Computed
+					},
+				},
+			},
+
+			// Computed
+			"state": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"time_created": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+			"time_updated": {
+				Type:     schema.TypeString,
+				Computed: true,
+			},
+		},
+	}
+}
+
+func createAlarm(d *schema.ResourceData, m interface{}) error {
+	sync := &AlarmResourceCrud{}
+	sync.D = d
+	sync.Client = m.(*OracleClients).monitoringClient
+
+	return CreateResource(d, sync)
+}
+
+func readAlarm(d *schema.ResourceData, m interface{}) error {
+	sync := &AlarmResourceCrud{}
+	sync.D = d
+	sync.Client = m.(*OracleClients).monitoringClient
+
+	return ReadResource(sync)
+}
+
+func updateAlarm(d *schema.ResourceData, m interface{}) error {
+	sync := &AlarmResourceCrud{}
+	sync.D = d
+	sync.Client = m.(*OracleClients).monitoringClient
+
+	return UpdateResource(d, sync)
+}
+
+func deleteAlarm(d *schema.ResourceData, m interface{}) error {
+	sync := &AlarmResourceCrud{}
+	sync.D = d
+	sync.Client = m.(*OracleClients).monitoringClient
+	sync.DisableNotFoundRetries = true
+
+	return DeleteResource(d, sync)
+}
+
+type AlarmResourceCrud struct {
+	BaseCrud
+	Client                 *oci_monitoring.MonitoringClient
+	Res                    *oci_monitoring.Alarm
+	DisableNotFoundRetries bool
+}
+
+func (s *AlarmResourceCrud) ID() string {
+	return *s.Res.Id
+}
+
+func (s *AlarmResourceCrud) CreatedPending() []string {
+	return []string{}
+}
+
+func (s *AlarmResourceCrud) CreatedTarget() []string {
+	return []string{
+		string(oci_monitoring.AlarmLifecycleStateActive),
+	}
+}
+
+func (s *AlarmResourceCrud) DeletedPending() []string {
+	return []string{
+		string(oci_monitoring.AlarmLifecycleStateDeleting),
+	}
+}
+
+func (s *AlarmResourceCrud) DeletedTarget() []string {
+	return []string{
+		string(oci_monitoring.AlarmLifecycleStateDeleted),
+	}
+}
+
+func (s *AlarmResourceCrud) Create() error {
+	request := oci_monitoring.CreateAlarmRequest{}
+
+	if body, ok := s.D.GetOkExists("body"); ok {
+		tmp := body.(string)
+		request.Body = &tmp
+	}
+
+	if compartmentId, ok := s.D.GetOkExists("compartment_id"); ok {
+		tmp := compartmentId.(string)
+		request.CompartmentId = &tmp
+	}
+
+	if definedTags, ok := s.D.GetOkExists("defined_tags"); ok {
+		convertedDefinedTags, err := mapToDefinedTags(definedTags.(map[string]interface{}))
+		if err != nil {
+			return err
+		}
+		request.DefinedTags = convertedDefinedTags
+	}
+
+	request.Destinations = []string{}
+	if destinations, ok := s.D.GetOkExists("destinations"); ok {
+		interfaces := destinations.([]interface{})
+		tmp := make([]string, len(interfaces))
+		for i := range interfaces {
+			if interfaces[i] != nil {
+				tmp[i] = interfaces[i].(string)
+			}
+		}
+		request.Destinations = tmp
+	}
+
+	if displayName, ok := s.D.GetOkExists("display_name"); ok {
+		tmp := displayName.(string)
+		request.DisplayName = &tmp
+	}
+
+	if freeformTags, ok := s.D.GetOkExists("freeform_tags"); ok {
+		request.FreeformTags = objectMapToStringMap(freeformTags.(map[string]interface{}))
+	}
+
+	if isEnabled, ok := s.D.GetOkExists("is_enabled"); ok {
+		tmp := isEnabled.(bool)
+		request.IsEnabled = &tmp
+	}
+
+	if metricCompartmentId, ok := s.D.GetOkExists("metric_compartment_id"); ok {
+		tmp := metricCompartmentId.(string)
+		request.MetricCompartmentId = &tmp
+	}
+
+	if metricCompartmentIdInSubtree, ok := s.D.GetOkExists("metric_compartment_id_in_subtree"); ok {
+		tmp := metricCompartmentIdInSubtree.(bool)
+		request.MetricCompartmentIdInSubtree = &tmp
+	}
+
+	if namespace, ok := s.D.GetOkExists("namespace"); ok {
+		tmp := namespace.(string)
+		request.Namespace = &tmp
+	}
+
+	if pendingDuration, ok := s.D.GetOkExists("pending_duration"); ok {
+		tmp := pendingDuration.(string)
+		request.PendingDuration = &tmp
+	}
+
+	if query, ok := s.D.GetOkExists("query"); ok {
+		tmp := query.(string)
+		request.Query = &tmp
+	}
+
+	if repeatNotificationDuration, ok := s.D.GetOkExists("repeat_notification_duration"); ok {
+		tmp := repeatNotificationDuration.(string)
+		request.RepeatNotificationDuration = &tmp
+	}
+
+	if resolution, ok := s.D.GetOkExists("resolution"); ok {
+		tmp := resolution.(string)
+		request.Resolution = &tmp
+	}
+
+	if severity, ok := s.D.GetOkExists("severity"); ok {
+		request.Severity = oci_monitoring.AlarmSeverityEnum(severity.(string))
+	}
+
+	if suppression, ok := s.D.GetOkExists("suppression"); ok {
+		if tmpList := suppression.([]interface{}); len(tmpList) > 0 {
+			fieldKeyFormat := fmt.Sprintf("%s.%d.%%s", "suppression", 0)
+			tmp, err := s.mapToSuppression(fieldKeyFormat)
+			if err != nil {
+				return err
+			}
+			request.Suppression = &tmp
+		}
+	}
+
+	request.RequestMetadata.RetryPolicy = getRetryPolicy(s.DisableNotFoundRetries, "monitoring")
+
+	response, err := s.Client.CreateAlarm(context.Background(), request)
+	if err != nil {
+		return err
+	}
+
+	s.Res = &response.Alarm
+	return nil
+}
+
+func (s *AlarmResourceCrud) Get() error {
+	request := oci_monitoring.GetAlarmRequest{}
+
+	tmp := s.D.Id()
+	request.AlarmId = &tmp
+
+	request.RequestMetadata.RetryPolicy = getRetryPolicy(s.DisableNotFoundRetries, "monitoring")
+
+	response, err := s.Client.GetAlarm(context.Background(), request)
+	if err != nil {
+		return err
+	}
+
+	s.Res = &response.Alarm
+	return nil
+}
+
+func (s *AlarmResourceCrud) Update() error {
+	request := oci_monitoring.UpdateAlarmRequest{}
+
+	tmp := s.D.Id()
+	request.AlarmId = &tmp
+
+	if body, ok := s.D.GetOkExists("body"); ok {
+		tmp := body.(string)
+		request.Body = &tmp
+	}
+
+	if compartmentId, ok := s.D.GetOkExists("compartment_id"); ok {
+		tmp := compartmentId.(string)
+		request.CompartmentId = &tmp
+	}
+
+	if definedTags, ok := s.D.GetOkExists("defined_tags"); ok {
+		convertedDefinedTags, err := mapToDefinedTags(definedTags.(map[string]interface{}))
+		if err != nil {
+			return err
+		}
+		request.DefinedTags = convertedDefinedTags
+	}
+
+	request.Destinations = []string{}
+	if destinations, ok := s.D.GetOkExists("destinations"); ok {
+		interfaces := destinations.([]interface{})
+		tmp := make([]string, len(interfaces))
+		for i := range interfaces {
+			if interfaces[i] != nil {
+				tmp[i] = interfaces[i].(string)
+			}
+		}
+		request.Destinations = tmp
+	}
+
+	if displayName, ok := s.D.GetOkExists("display_name"); ok {
+		tmp := displayName.(string)
+		request.DisplayName = &tmp
+	}
+
+	if freeformTags, ok := s.D.GetOkExists("freeform_tags"); ok {
+		request.FreeformTags = objectMapToStringMap(freeformTags.(map[string]interface{}))
+	}
+
+	if isEnabled, ok := s.D.GetOkExists("is_enabled"); ok {
+		tmp := isEnabled.(bool)
+		request.IsEnabled = &tmp
+	}
+
+	if metricCompartmentId, ok := s.D.GetOkExists("metric_compartment_id"); ok {
+		tmp := metricCompartmentId.(string)
+		request.MetricCompartmentId = &tmp
+	}
+
+	if metricCompartmentIdInSubtree, ok := s.D.GetOkExists("metric_compartment_id_in_subtree"); ok {
+		tmp := metricCompartmentIdInSubtree.(bool)
+		request.MetricCompartmentIdInSubtree = &tmp
+	}
+
+	if namespace, ok := s.D.GetOkExists("namespace"); ok {
+		tmp := namespace.(string)
+		request.Namespace = &tmp
+	}
+
+	if pendingDuration, ok := s.D.GetOkExists("pending_duration"); ok {
+		tmp := pendingDuration.(string)
+		request.PendingDuration = &tmp
+	}
+
+	if query, ok := s.D.GetOkExists("query"); ok {
+		tmp := query.(string)
+		request.Query = &tmp
+	}
+
+	if repeatNotificationDuration, ok := s.D.GetOkExists("repeat_notification_duration"); ok {
+		tmp := repeatNotificationDuration.(string)
+		request.RepeatNotificationDuration = &tmp
+	}
+
+	if resolution, ok := s.D.GetOkExists("resolution"); ok {
+		tmp := resolution.(string)
+		request.Resolution = &tmp
+	}
+
+	if severity, ok := s.D.GetOkExists("severity"); ok {
+		request.Severity = oci_monitoring.AlarmSeverityEnum(severity.(string))
+	}
+
+	if suppression, ok := s.D.GetOkExists("suppression"); ok {
+		if tmpList := suppression.([]interface{}); len(tmpList) > 0 {
+			fieldKeyFormat := fmt.Sprintf("%s.%d.%%s", "suppression", 0)
+			tmp, err := s.mapToSuppression(fieldKeyFormat)
+			if err != nil {
+				return err
+			}
+			request.Suppression = &tmp
+		}
+	}
+
+	request.RequestMetadata.RetryPolicy = getRetryPolicy(s.DisableNotFoundRetries, "monitoring")
+
+	response, err := s.Client.UpdateAlarm(context.Background(), request)
+	if err != nil {
+		return err
+	}
+
+	s.Res = &response.Alarm
+	return nil
+}
+
+func (s *AlarmResourceCrud) Delete() error {
+	request := oci_monitoring.DeleteAlarmRequest{}
+
+	tmp := s.D.Id()
+	request.AlarmId = &tmp
+
+	request.RequestMetadata.RetryPolicy = getRetryPolicy(s.DisableNotFoundRetries, "monitoring")
+
+	_, err := s.Client.DeleteAlarm(context.Background(), request)
+	return err
+}
+
+func (s *AlarmResourceCrud) SetData() error {
+	if s.Res.Body != nil {
+		s.D.Set("body", *s.Res.Body)
+	}
+
+	if s.Res.CompartmentId != nil {
+		s.D.Set("compartment_id", *s.Res.CompartmentId)
+	}
+
+	if s.Res.DefinedTags != nil {
+		s.D.Set("defined_tags", definedTagsToMap(s.Res.DefinedTags))
+	}
+
+	s.D.Set("destinations", s.Res.Destinations)
+
+	if s.Res.DisplayName != nil {
+		s.D.Set("display_name", *s.Res.DisplayName)
+	}
+
+	s.D.Set("freeform_tags", s.Res.FreeformTags)
+
+	if s.Res.IsEnabled != nil {
+		s.D.Set("is_enabled", *s.Res.IsEnabled)
+	}
+
+	if s.Res.MetricCompartmentId != nil {
+		s.D.Set("metric_compartment_id", *s.Res.MetricCompartmentId)
+	}
+
+	if s.Res.MetricCompartmentIdInSubtree != nil {
+		s.D.Set("metric_compartment_id_in_subtree", *s.Res.MetricCompartmentIdInSubtree)
+	}
+
+	if s.Res.Namespace != nil {
+		s.D.Set("namespace", *s.Res.Namespace)
+	}
+
+	if s.Res.PendingDuration != nil {
+		s.D.Set("pending_duration", *s.Res.PendingDuration)
+	}
+
+	if s.Res.Query != nil {
+		s.D.Set("query", *s.Res.Query)
+	}
+
+	if s.Res.RepeatNotificationDuration != nil {
+		s.D.Set("repeat_notification_duration", *s.Res.RepeatNotificationDuration)
+	}
+
+	if s.Res.Resolution != nil {
+		s.D.Set("resolution", *s.Res.Resolution)
+	}
+
+	s.D.Set("severity", s.Res.Severity)
+
+	s.D.Set("state", s.Res.LifecycleState)
+
+	if s.Res.Suppression != nil {
+		s.D.Set("suppression", []interface{}{SuppressionToMap(s.Res.Suppression)})
+	} else {
+		s.D.Set("suppression", nil)
+	}
+
+	if s.Res.TimeCreated != nil {
+		s.D.Set("time_created", s.Res.TimeCreated.String())
+	}
+
+	if s.Res.TimeUpdated != nil {
+		s.D.Set("time_updated", s.Res.TimeUpdated.String())
+	}
+
+	return nil
+}
+
+func (s *AlarmResourceCrud) mapToSuppression(fieldKeyFormat string) (oci_monitoring.Suppression, error) {
+	result := oci_monitoring.Suppression{}
+
+	if description, ok := s.D.GetOkExists(fmt.Sprintf(fieldKeyFormat, "description")); ok {
+		tmp := description.(string)
+		result.Description = &tmp
+	}
+
+	if timeSuppressFrom, ok := s.D.GetOkExists(fmt.Sprintf(fieldKeyFormat, "time_suppress_from")); ok {
+		tmp, err := time.Parse(time.RFC3339, timeSuppressFrom.(string))
+		if err != nil {
+			return result, err
+		}
+		result.TimeSuppressFrom = &oci_common.SDKTime{Time: tmp}
+	}
+
+	if timeSuppressUntil, ok := s.D.GetOkExists(fmt.Sprintf(fieldKeyFormat, "time_suppress_until")); ok {
+		tmp, err := time.Parse(time.RFC3339, timeSuppressUntil.(string))
+		if err != nil {
+			return result, err
+		}
+		result.TimeSuppressUntil = &oci_common.SDKTime{Time: tmp}
+	}
+
+	return result, nil
+}
+
+func SuppressionToMap(obj *oci_monitoring.Suppression) map[string]interface{} {
+	result := map[string]interface{}{}
+
+	if obj.Description != nil {
+		result["description"] = string(*obj.Description)
+	}
+
+	if obj.TimeSuppressFrom != nil {
+		result["time_suppress_from"] = obj.TimeSuppressFrom.Format(time.RFC3339Nano)
+	}
+
+	if obj.TimeSuppressUntil != nil {
+		result["time_suppress_until"] = obj.TimeSuppressUntil.Format(time.RFC3339Nano)
+	}
+
+	return result
+}